`qHash` for `TNullableValue<T>` returned `unsigned int`, silently
truncating Qt6's 64-bit `size_t` hash to 32 bits. On a 64-bit build,
`qHash(TNullableValue<QString>("a"), 17U)` produced `1848887763` instead
of the expected `5172881913348868563`.

## Change

- **`Include/istd/TNullableValue.h`**: Updated the `qHash` free function
template to use `size_t` for both the return type and `seed` parameter,
matching Qt6's `qHash` signature.

```cpp
// Before
inline unsigned int qHash(const istd::TNullableValue<T> &key, unsigned int seed)

// After
inline size_t qHash(const istd::TNullableValue<T> &key, size_t seed)
```

+++ Include/istd/TNullableValue.h
@@ -410,6 +410,6 @@ bool TNullableValue<T>::operator!=(const T& value) const
 
 
 template <class T>
-inline unsigned int qHash(const istd::TNullableValue<T> &key, unsigned int seed){
+inline size_t qHash(const istd::TNullableValue<T> &key, size_t seed){
 	return qHash(*key, seed);
 }
